+++ src/main/java/racingcar/domain/car/power/Power.java
@@ -1,4 +1,4 @@
-package racingcar.domain.car;
+package racingcar.domain.car.power;
 
 @FunctionalInterface
 public interface Power {
