+++ src/main/java/racingcar/domain/game/GameResult.java
@@ -1,6 +1,5 @@
-package racingcar.domain;
+package racingcar.domain.game;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class GameResult {
