+++ src/test/java/racingcar/domain/car/ParticipateCarsTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import racingcar.domain.car.power.FixedPower;
 
 import java.util.Arrays;
 import java.util.List;
